+++ cJSON.h
@@ -105,10 +105,10 @@ extern cJSON *cJSON_CreateArray(void);
 extern cJSON *cJSON_CreateObject(void);
 
 /* These utilities create an Array of count items. */
-extern cJSON *cJSON_CreateIntArray(const int *numbers, int count);
-extern cJSON *cJSON_CreateFloatArray(const float *numbers, int count);
-extern cJSON *cJSON_CreateDoubleArray(const double *numbers, int count);
-extern cJSON *cJSON_CreateStringArray(const char **strings, int count);
+extern cJSON *cJSON_CreateIntArray(const int *numbers, size_t count);
+extern cJSON *cJSON_CreateFloatArray(const float *numbers, size_t count);
+extern cJSON *cJSON_CreateDoubleArray(const double *numbers, size_t count);
+extern cJSON *cJSON_CreateStringArray(const char **strings, size_t count);
 
 /* Append item to the specified array/object. */
 extern void cJSON_AddItemToArray(cJSON *array, cJSON *item);
